accounts/obtain_access_token.py
from rest_framework.authtoken.views import ObtainAuthToken
from rest_framework.authtoken.models import Token
from rest_framework.response import Response
from rest_framework import status
from django.core.validators import validate_email
from django.core.exceptions import ValidationError
from django.shortcuts import get_object_or_404
from .models import User

class CustomObtainAuthToken(ObtainAuthToken):
    def post(self, request, *args, **kwargs):
        serializer = self.get_serializer(data=request.data)

        serializer.is_valid(raise_exception=True)
        user = serializer.validated_data['user']
        
        if (request.data.get("refresh", False)):
            token = Token.objects.get(user=user)
            if (token):
                token.delete()
            token = Token.objects.create(user=user)
        else:
            token, created = Token.objects.get_or_create(user=user)
        return Response({
            'token': token.key,
            'email': user.email,
            'first_name': user.first_name,
            'last_name': user.last_name,
            'is_admin': user.is_admin,
            'date_joined': user.date_joined,
            'is_active': user.is_active,
        })
    
obtain_auth_token = CustomObtainAuthToken.as_view()