requirements.txt
asgiref==3.6.0
asttokens==2.2.1
autopep8==2.0.2
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
decorator==5.1.1
defusedxml==0.7.1
Django==4.2
django-cors-headers==3.14.0
django-filter==23.1
djangorestframework==3.14.0
executing==1.2.0
httpie==3.2.1
idna==3.4
ipython==8.12.0
jedi==0.18.2
markdown-it-py==2.2.0
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.0.4
numpy==1.24.2
pandas==2.0.0
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.38
pure-eval==0.2.2
pycodestyle==2.10.0
Pygments==2.15.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
requests-toolbelt==0.10.1
rich==13.3.4
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tqdm==4.65.0
traitlets==5.9.0
tzdata==2023.3
urllib3==1.26.15
wcwidth==0.2.6
